# Conflicts:
#	ahcyworld-app/src/webpage/components/Header/Header.jsx
#	ahcyworld-app/src/webpage/components/mybox/MyBox.css
#	ahcyworld-app/src/webpage/components/mybox/MyBox.jsx
#	ahcyworld-app/src/webpage/pages/Home.jsx

+++ ahcyworld-app/src/minihompy/components/post/utils/commentUtils.js
@@ -1,27 +1,43 @@
 import { CommentAction } from "../../../../redux/actions/CommentAction";
-
+import * as Swal from "../../../../apis/alert"
 export const commentListAxios = async (dispatch, postId, postName) => {
   try {
     await dispatch(CommentAction.commentListAxios(postId,postName));
   } catch (e) {
-    alert(e);
+    Swal.alert("댓글정보를 불러오는데 실패했습니다",e,"error")
+    
   }
 };
 
 export const commentWriteAxios = async (dispatch, userInfo, post, content, postName) => {
   try {
     await dispatch(CommentAction.commentWriteAxios(userInfo, post, content,postName));
+    Swal.alert("댓글작성을 성공했습니다.","작성완료","success")
   } catch (e) {
-    alert(e);
+    Swal.alert("댓글작성을 실패했습니다",e,"error")
   }
 };
 
 export const commentDeleteAxios = async (dispatch, commentId, postId,postName) =>{
+  
+  
   try{
+
+    // const confuse = await Swal.confirm('삭제하시겠습니까?','다시한번 확인해주세요','warning');
+    // console.log(confuse)
+
+    // if(!confuse) {
+    //   console.log('??',confuse)
+    //   return;
+    // }
+      
+
+    if(!window.confirm('삭제하시겠습니까?')) return;
+
     await dispatch(CommentAction.commentDeleteAxios(commentId, postId, postName))
-    
+    Swal.alert("댓글 삭제 완료.","삭제완료","success")
   }catch(e){
-    alert(e)
+    Swal.alert("댓글삭제를 실패했습니다",e,"error")
   }
 } 
 
@@ -33,8 +49,6 @@ export const showState =(dispatch,state) =>{
   }else if(state === true){
     data =false
   }
-
-  console.log('first',data)
 
   dispatch(CommentAction.showState(data))
 }
